Got around to generating the frustum planes
Not sure if they work or not though, since everything seems to have broken during the move from VS2015 to VS2017. Data renders everywhere EXCEPT the target window, and it's untestable without having valid data (i.e. a working version of the same program) to check against during debugging. I'll have to set up two simultaneous versions tomorrow and see if they diverge anywhere. If they don't, then I'll try to change the target windows version and see if it changes anything; if that doesn't work, then I'll carefully back-up-and-revert my way towards a working system :)

+++ AthruRewrite/Athru/Athru/Chunk.cpp
@@ -14,7 +14,11 @@ Chunk::Chunk(byteSigned offsetFromHomeX, byteSigned offsetFromHomeZ)
 	for (eightByteSigned i = 0; i < CHUNK_VOLUME; i += 1)
 	{
 		chunkBoxecules[i] = new Boxecule(0.7f, 0.5f, 0.5f, 0.8f);
-		DirectX::XMVECTOR boxeculePos = _mm_set_ps(1, ((i / CHUNK_WIDTH) % CHUNK_WIDTH) + offsetFromHomeZ, i % CHUNK_WIDTH, i / (CHUNK_WIDTH * CHUNK_WIDTH) + offsetFromHomeX);
+
+		DirectX::XMVECTOR boxeculePos = _mm_set_ps(1, ((float)((i / CHUNK_WIDTH) % CHUNK_WIDTH)) + offsetFromHomeZ,
+												       (float)(i % CHUNK_WIDTH),
+													   (float)(i / (CHUNK_WIDTH * CHUNK_WIDTH)) + offsetFromHomeX);
+
 		DirectX::XMVECTOR boxeculeRot = DirectX::XMQuaternionRotationRollPitchYaw(0, 0, 0);
 		chunkBoxecules[i]->FetchTransformations() = SQT(boxeculeRot, boxeculePos, 1);
  	}
